feat:: 토큰처리 위치 변경
