+++ src/app/layout.tsx
@@ -1,9 +1,12 @@
+// app/layout.tsx
 import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 import './globals.css';
 import { ReactQueryProvider } from '@/providers/ReactQueryProvider';
 import MobileFooter from '@/components/layout/footer/MobileFooter';
+import { getAccessToken } from '../lib/getTokenFromCookies';
 import '../lib/fetchToken';
+import { TokenHandler } from '@/components/TokenHandler';
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
@@ -20,19 +23,15 @@ export const metadata: Metadata = {
   description: 'Generated by create next app',
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  // if (process.env.NODE_ENV === 'development') {
-  //   init();
-  // }
+}: Readonly<{ children: React.ReactNode }>) {
+  const accessToken = await getAccessToken();
 
   return (
     <html lang="en">
-      {/* <MswComponent /> */}
       <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
+        <TokenHandler accessToken={accessToken} />
         <ReactQueryProvider>{children}</ReactQueryProvider>
         <MobileFooter />
       </body>
